+++ packages/edition-slices/src/slices/secondaryone/styles.js
@@ -1,0 +1,9 @@
+import { spacing } from "@times-components/styleguide";
+
+const styles = {
+  container: {
+    paddingBottom: spacing(2)
+  }
+};
+
+export default styles;