+++ packages/edition-slices/src/tiles/tile-u/styles/index.js
@@ -1,0 +1,23 @@
+import { fonts, spacing } from "@times-components/styleguide";
+
+const styles = {
+  container: {
+    flexDirection: "row",
+    padding: spacing(2)
+  },
+  headline: {
+    fontFamily: fonts.headline,
+    fontSize: 25,
+    lineHeight: 25,
+    paddingVertical: spacing(1)
+  },
+  imageContainer: {
+    width: "66.66%"
+  },
+  summaryContainer: {
+    paddingVertical: spacing(1),
+    width: "33.33%"
+  }
+};
+
+export default styles;