+++ packages/slice-layout/src/templates/leadoneandone/proptypes.js
@@ -1,8 +1,14 @@
 import PropTypes from "prop-types";
+import { editionBreakpoints } from "@times-components/styleguide";
 
 const propTypes = {
+  breakpoint: PropTypes.string,
   renderLead: PropTypes.func.isRequired,
   renderSupport: PropTypes.func.isRequired
 };
 
-export default propTypes;
+const defaultProps = {
+  breakpoint: editionBreakpoints.small
+};
+
+export { propTypes, defaultProps };