+++ packages/slice-layout/src/templates/leadoneandone/styles.js
@@ -1,0 +1,14 @@
+const styles = {
+  container: {
+    flex: 1,
+    flexDirection: "row"
+  },
+  leadItem: {
+    width: "70%"
+  },
+  supportItem: {
+    width: "30%"
+  }
+};
+
+export default styles;